Explicitly cast a l_ulong to an unsigned long to make all arch's happy with the printf format.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@113579 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/compat/linux/linux_misc.c
@@ -878,7 +878,7 @@ linux_personality(struct thread *td, struct linux_personality_args *args)
 {
 #ifdef DEBUG
 	if (ldebug(personality))
-		printf(ARGS(personality, "%ld"), args->per);
+		printf(ARGS(personality, "%lu"), (unsigned long)args->per);
 #endif
 #ifndef __alpha__
 	if (args->per != 0)
